Split app.py
Functions contains two generate functions, one for background samples and one for spectra. Both call the same __process_spectra() function.
